% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfordActivity.R
\name{binfordActivity}
\alias{binfordActivity}
\title{binfordActivity}
\usage{
binfordActivity(d, LakeName = "Lake Name", live = 80000, PbBk = 0, BiBk = 0, 
CsBk = 0, PbBkErr = 0, BiBkErr = 0, CsBkErr = 0, graph.units="dpm")
}
\arguments{
\item{d}{the matrix or data frame containing the input data.}

\item{LakeName}{Name of the lake being analyzed. Defaults to "Lake Name".}

\item{live}{Time in seconds that the samples were counted for. Defaults to 80000.}

\item{PbBk}{Background 210Pb (cpm). Defaults to 0.}

\item{BiBk}{Background 214Bi (cpm). Defaults to 0.}

\item{CsBk}{Background 137Cs (cpm). Defaults to 0.}

\item{PbBkErr}{Background Error for 210Pb (cpm). Defaults to 0.}

\item{BiBkErr}{Background Error for 214Bi (cpm). Defaults to 0.}

\item{CsBkErr}{Background Error for 137Cs (cpm). Defaults to 0.}

\item{graph.units}{determines whether activities should be plotted as dpm/g ("dpm") or Bq/g ("bq"). Defaults to "dpm".}
}
\value{
\code{binfordActivityOutput} table of results for use with the \code{binfordDates} function.
}
\description{
\code{binfordActivity} uses the appropriate Region of Interest (ROI) values from the gamma counter to correct the 210Pb, 214Bi and 137Cs activities for the amount of sediment in the gamma tube as well as the background measurements of the gamma counter.
}
\details{
\code{binfordActivity} calculates the activity of the 210Pb, 137Cs, and 214Bi isotopes, corrected for the efficiency of the gamma counter device, and the time between sediment core collection and analysis. A plot is generated of the corrected isotope activities vs. sediment core depth. The data to be examined should be in the form of a matrix or data frame with the following 14 columns:
\itemize{
 \item INTTOP: top of the analyzed interval (cm)
 \item INTBOT: bottom of the analyzed interval (cm)
 \item DATECORR: number of days between core collection and analysis of the interval
 \item ROI1: sum of the activity of the 5 channel region of interest immediately less than the 210Pb photopeak (ROI1 from the Maestro ROI Report)
 \item ROI2: sum of the activity of the 10 channel region of interest centered on the 210Pb photopeak (45.5keV; ROI2 from the Maestro ROI Report)
 \item ROI3: sum of the activity of the 5 channel region of interest immediately greater than the 210Pb photopeak (ROI3 from the Maestro ROI Report)
 \item ROI4: sum of the activity of the 5 channel region of interest immediately less than the 214Bi photopeak (ROI10 from the Maestro ROI Report)
 \item ROI5: sum of the activity of the 10 channel region of interest centered on the 214Bi photopeak (609keV; ROI11 from the Maestro ROI Report)
 \item ROI6: sum of the activity of the 5 channel region of interest immediately greater than the 214Bi photopeak (ROI12 from the Maestro ROI Report)
 \item ROI7: sum of the activity of the 5 channel region of interest immediately less than the 137Cs photopeak (ROI13 from the Maestro ROI Report)
 \item ROI8: sum of the activity of the 10 channel region of interest centered on the 137Cs photopeak (662keV; ROI14 from the Maestro ROI Report)
 \item ROI9: sum of the activity of the 5 channel region of interest immediately greater than the 137Cs photopeak (ROI15 from the Maestro ROI Report)
 \item WTinTube: mass (g) of dried sediment added to the gamma tube (equivalent to WT_GAMMA in \code{binfordRho})
 \item HTinTube: height (mm) of the sediment added to the gamma tube, generally measured to 3 decimal places using an accurate set of calipers
 }

NOTE: Columns 1 and 2 must have values for all samples prepared for gamma analysis (the same values entered into Columns 1 and 2 of \code{binfordRho}), but no values for columns 3-9 in cases where samples have not yet been run or are not required (in cases where background has been reached).

\code{binfordActivity} is intended to be run in an iterative function following analysis of each sample, thereby allowing accurate and timely determination of when supported levels of 210Pb have been reached, in order to avoid the superfluous analysis of samples beyond background.

It is very important to use the correct background and error measurements (e.g. PbBk, PbBkErr, etc.). That is, the values specific to both the gamma counter used and the time period when the samples were measured.
}
\examples{
#Imports the example activity data for a sediment core subsampled prior to freeze-drying
data(binfordActivityInput)
binfordActivity(binfordActivityInput)
}
\references{
Schelske CL, Peplow A, Brenner M, Spencer CN (1994) Low-background gamma counting: applications for 210Pb dating of sediments Journal of Paleolimnology 10: 115-128
}
\author{
Adam Jeziorski, Joshua Thienpont
}
