% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfordActivityInput.R
\docType{data}
\name{binfordActivityInput}
\alias{binfordActivityInput}
\title{An example input file for \code{binfordActivity}}
\format{a data frame containing 25 observations of 14 variables}
\usage{
binfordActivityInput
}
\description{
This dataset demonstrates the required input format for \code{binfordActivity}.
The \code{binfordActivityInput} data frame contains the following columns: 
\itemize{
 \item INTTOP: top of the analyzed interval (cm)
 \item INTBOT: bottom of the analyzed interval (cm)
 \item DATECORR: number of days between core collection and analysis of the sample
 \item ROI1: sum of the activity of the 5 channel region of interest immediately less than the 210Pb photopeak (ROI1 from the Maestro ROI Report)
 \item ROI2: sum of the activity of the 10 channel region of interest centered on the 210Pb photopeak (45.5keV; ROI2 from the Maestro ROI Report)
 \item ROI3: sum of the activity of the 5 channel region of interest immediately greater than the 210Pb photopeak (ROI3 from the Maestro ROI Report)
 \item ROI4: sum of the activity of the 5 channel region of interest immediately less than the 214Bi photopeak (ROI10 from the Maestro ROI Report)
 \item ROI5: sum of the activity of the 10 channel region of interest centered on the 214Bi photopeak (609keV; ROI11 from the Maestro ROI Report)
 \item ROI6: sum of the activity of the 5 channel region of interest immediately greater than the 214Bi photopeak (ROI12 from the Maestro ROI Report)
 \item ROI7: sum of the activity of the 5 channel region of interest immediately less than the 137Cs photopeak (ROI13 from the Maestro ROI Report)
 \item ROI8: sum of the activity of the 10 channel region of interest centered on the 137Cs photopeak (662keV; ROI14 from the Maestro ROI Report)
 \item ROI9: sum of the activity of the 5 channel region of interest immediately greater than the 137Cs photopeak (ROI15 from the Maestro ROI Report)
 \item WTinTube: mass (g) of dried sediment added to the gamma tube (equivalent to WT_GAMMA in \code{binfordRho})
 \item HTinTube: height (mm) of the sediment added to the gamma tube, generally measured to 3 decimal places using an accurate set of calipers
 }
}
\author{
Adam Jeziorski
}
