% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfordDates.R
\name{binfordDates}
\alias{binfordDates}
\title{binfordDates}
\usage{
binfordDates(binfordRhoOutput, binfordActivityOutput, LakeName = "Lake Name", 
CoreName = "KB1", CoreDate = 2006, CIC = FALSE)
}
\arguments{
\item{binfordRhoOutput}{matrix or data frame containing the output data from \code{binfordRho}}

\item{binfordActivityOutput}{matrix or data frame containing the output data from \code{binfordActivity}}

\item{LakeName}{(defaults to "Lake Name"): name of the lake being analyzed}

\item{CoreName}{(defaults to "KB1"): name of the core being analyzed}

\item{CoreDate}{(defaults to 2006): year of core collection}

\item{CIC}{(defaults to FALSE): logical to control whether to also calculate and plot CIC dates (NOTE: this was still being worked on when the functions were rendered obsolete, and remains incomplete)}
}
\value{
\code{binfordDatesOutput} returns a table with 15 columns:
\itemize{
 \item IntTop: depth (cm) to the top of the analyzed interval
 \item IntBot: depth (cm) to the bottom of the analyzed interval
 \item IntMid: depth (cm) to the midpoint of the analyzed interval
 \item TTop: age (years) at the top of the interval
 \item SDTTop: standard error of the age at the top of the interval
 \item TBot: age (years) at the bottom of the interval.
 \item SDTBot: standard error of the age at the bottom of the interval
 \item SedRate: sedimentation rate (g/cm^2/yr).
 \item SDSedRate: standard error of the sedimentation rate
 \item SumTop: cumulative dry mass from the bottom of the core to the top of the interval
 \item DTop: date at the top of the interval
 \item DBot: date at the bottom of the interval
 \item DMid: date at the midpoint of the interval
 \item CICDMid: CIC Date at the midpoint of the interval
 \item OrgSedRate: organic sedimentation rate
 }
}
\description{
\code{binfordDates} uses the output of \code{binfordRho} and \code{binfordActivity} to calculate 210Pb dates using the Constant Rate of Supply (CRS) model.
}
\details{
\code{binfordDates} determines the unsupported fraction of 210Pb activity using the procedure outlined in Binford (1990) and then calculates dates based on the constant rate of supply model. For input, it requires the output data from \code{binfordRho} and \code{binfordActivity}.
}
\examples{
#Import example input data
data(binfordRhoInput)
data(binfordActivityInput)
binfordDates(binfordRho(binfordRhoInput), binfordActivity(binfordActivityInput), 
LakeName="Test", CoreName="KB1", CIC=TRUE, CoreDate=2010)

}
\references{
Binford MW (1990) Calculation and uncertainty analysis of 210PB dates for PIRLA project lake sediment cores. Journal of Paleolimnology 3: 253-267
}
\author{
Adam Jeziorski, Joshua Thienpont
}
