% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfordRho.R
\name{binfordRho}
\alias{binfordRho}
\title{binfordRho}
\usage{
binfordRho(d, bagwt=3.25, subsample=TRUE, coreD=7.62)
}
\arguments{
\item{d}{matrix or data frame containing the input data}

\item{bagwt}{(defaults to 3.25): average weight (g) of your brand/size of "Whirlpak" bags}

\item{subsample}{(Defaults to TRUE): logical to indicate whether the samples were subsampled before freeze-drying}

\item{coreD}{(defaults to 7.62):inner tube diameter (cm) of the sediment corer used to collect the core}
}
\value{
\code{binfordRhoOutput} table of results for use with the \code{binfordDates} function.
}
\description{
\code{binfordRho} calculates the percent water and dry sediment density for the individual samples of a sediment core freeze-dried in preparation for gamma dating. The values for intervening (i.e. non-prepared samples) are interpolated.
}
\details{
\code{binfordRho} interpolates the values of dry sediment density needed for dating a sediment core from the dry masses of the freeze-dried intervals and the bag weights of wet sediment (by calculating and interpolating percent water). The data to be examined should be in the form of a matrix or data frame with the following 7 columns:
\itemize{
 \item INTTOP: depth (cm) to the top of the analyzed interval
 \item INTBOT: depth (cm) to the bottom of the analyzed interval
 \item WT_SED+BAG: mass (g) of the wet sediment and the WhirlPak bag
 \item WT_VIAL: mass (g) of the empty vial used for freeze drying (disregarded if \code{subsample}=FALSE)
 \item WT_VIAL+SED: mass (g) of the vial after adding the subsample of wet sediment (disregarded if \env{subsample}=FALSE)
 \item WT_FD = mass (g) of the vial of sediment after freeze drying (if \code{subsample}=FALSE, this value should instead be the mass (g) of the dry sediment and the Whirlpak bag)
 \item WT_GAMMA: mass (g) of freeze-dried sediment added to the gamma tube
 }

NOTE: Columns 1-3 must have values for the entire core length. 
Columns 4-7 will only have values for the intervals prepared for dating by freeze-drying, with the values for non-prepared intervals left blank.
If the entire "Whirlpak" bag was freeze-dried (i.e. subsample=FALSE), then Columns 4-5 should be left blank and the mass (g) of the bag of sediment after freeze drying entered into Column 6.
}
\examples{
#Imports the example rho data for a sediment core subsampled prior to freeze-drying
data(binfordRhoInput)
binfordRho(binfordRhoInput)
}
\author{
Adam Jeziorski, Joshua Thienpont
}
