% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirInfo.R
\docType{data}
\name{chirInfo}
\alias{chirInfo}
\title{Chironomid ecological information}
\format{
A data frame containing 211 observations of 18 variables.
}
\usage{
chirInfo
}
\description{
\code{chirInfo} is a data frame tabulating several aspects of chironomid ecology into a single file. 

The \code{chirInfo} data frame contains the following 18 columns:
\itemize{
  \item Column 1: taxon name
  \item Columns 2-5: taxonomic information
  \item Columns 6-11: ecological info taken from Brooks (2007)
  \item Columns 12-15: ecological info taken from Coffman (1978)
  \item Column 16: taxonomic codes used in Roberto Quinlan's PhD thesis and the 2001/2010 papers
  \item Columns 17-18: optima/tolerances from the Quinlan 2010 VWHO model
  }
}
\references{
Brooks SJ, Langdon PG, Heiri O (2007) The identification and use of Palaearctic Chironomidae larvae in paleoecology. Quat Res (London): Technical Guide no. 10.

Coffman WP (1978) Chironomidae. pp. 345-376. In: Merrit, RW, Cummins KW (Eds). An introduction of aquatic  insects  of  North  America.  Kendall  Hunt Publishing, Dubuque, 441p

Quinlan R, Smol JP (2001) Chironomid-based inference models for estimating end-of-summer hypolimnetic oxygen from south-central Ontario shield lakes. Freshwater Biology 46: 1529-1551

Quinlan R, Smol JP (2010) Use of Chaoborus subfossil mandibles in models for inferring past hypolimnetic oxygen. Journal of Paleolimnology 44: 43-50
}
