% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladCount.R
\name{cladCount}
\alias{cladCount}
\title{cladCount}
\usage{
cladCount(d, percCutoff=2, sampleCutoff=2, outputType="indiv")
}
\arguments{
\item{d}{matrix or data frame containing the input data}

\item{percCutoff}{(defaults to 2): minimum relative abundance (i.e. \%) required for a taxon to be included in the the reduced subset ('gt' - abbreviation of 'greater than')}

\item{sampleCutoff}{(defaults to 2): minimum number of samples a taxon must be present in with at least a relative abundance of \code{percCutoff} for inclusion in the reduced subset ('gt')}

\item{outputType}{(defaults to 'indiv'): the format of the output, either individuals ('indiv'), relative abundance ('perc'), or the reduced subset ('gt')}
}
\value{
\code{cladCount} will return one of: 
\itemize{
 \item \code{indiv}: table listing the minimum number of individuals of each taxa represented in each interval
 \item \code{perc}: table listing the relative abundance of each taxa within each interval
 \item \code{gt}: relative abundances of those taxa meeting the \code{sampleCutoff} and \code{percCutoff} criteria
 }
}
\description{
\code{cladCount} determines the maximum number of individuals represented in raw counts of cladoceran subfossils
}
\details{
The input data should be in the form of a matrix or data frame with at least 3 columns:
\itemize{
  \item Column 1: taxon name
  \item Column 2: subfossil name
  \item Column 3+: count of each taxon/subfossil remain type
  }
NOTE: Special characters in the column headings may interfere with import of the input file. The script uses string matching to identify body parts, anything other than the following list in the body part column (i.e. spelling errors) will be ignored.
\itemize{
  \item Body part (number of subfossils per individual)
  \item Headshield (1)
  \item Carapace (2)
  \item PA Claw (2)
  \item Postabdomen (1)
  \item Mandible (2)
  \item Caudal Furca (2)
  \item Exopodite Segment (2)
  \item Basal Exp Segment (2)
  \item Exp Segment 2 (2)
  \item Exp Segment 3 (2)
  \item Antennule (2)
  \item Tail Stem (1)     
  }
}
\examples{
#load example cladoceran count data
data(cladCountInput)
cladCount(cladCountInput)

#Return the values for only taxa with greater than 4 percent abundance in at least 2 samples
cladCount(cladCountInput, percCutoff=4, sampleCutoff=2)

}
\references{
Korhola A, Rautio M (2001) 2. Cladocera and other branchiopod crustaceans. In: Smol JP, Birks HJB, Last WM (eds.) Tracking Environmental Change Using Lake Sediments. Volume 4: Zoological Indicators. Kluwer Academic Publishers, Dordrecht, The Netherlands, pp 4-41
t' increases by 1 for each new taxa
}
\author{
Adam Jeziorski, Andrew Labaj
}
