% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladCountInput.R
\docType{data}
\name{cladCountInput}
\alias{cladCountInput}
\title{An example input file for cladCount}
\format{
a data frame containing 106 observations of 27 variables.
}
\usage{
cladCountInput
}
\description{
This dataset demonstrates the required input format for \code{cladCount}
The \code{cladCountInput} data frame contains the following columns:
\itemize{
  \item Column 1: taxon name
  \item Column 2: subfossil name
  \item Column 3+: count of each taxon/subfossil remain type
  }
}
\author{
Adam Jeziorski
}
