% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipteranVWHO.R
\name{dipteranVWHO}
\alias{dipteranVWHO}
\title{dipteranVWHO}
\usage{
dipteranVWHO(d, evaluate=FALSE, percentileCut=5, lowCount=50)
}
\arguments{
\item{d}{matrix or data frame containing the input data}

\item{evaluate}{(defaults to FALSE): logical to indicate whether to display the model information}

\item{percentileCut}{(defaults to 5): cutoff value for flagging samples in the plots as having poor modern analogs}

\item{lowCount}{(defaults to 50): cutoff value to flag samples that have low counts (i.e. less than 50 individuals)}
}
\value{
\code{output} list  of 4 sublists 
\itemize{
  \item formattedData contains the sample info, the raw chironomid counts, the raw chaoborid counts, the chironomid counts aggregated by taxon code to match the taxonomy used in vwhoQuinlan 2010 dataset, and vector of the 44 taxon codes used in the model.
  \item vwhoModel the model object constructed by the 'WA' function provided by 'rioja'
  \item vwhoResults contains model ouput, relative abundance of the dipterans used in the model, relative abundances of all dipterans in the dataset
  \item analogResults contains the output from the analog analyis, the close modern analogs, and the minimum dissimilarity values
  }
Three plots are produced: minimum dissimilarity between each sample interval and the calibration set data, sediment depth vs dipteran-inferred VWHO, and a composite combining data from both graphs.
}
\description{
\code{dipteranVWHO} is a function to harmonize appropriately formatted chironomid count data with the calibration set described in Quinlan and Smol (2001, 2010), construct the described VWHO model, use the model to infer VWHO from the chironomid sample data, and finally perform an analog matching analysis between the sample data and the calibration set (that follows the procedures described in the 'analogue_methods' vignette provided by the 'analogue' package).
}
\details{
The required format of the input data is provided in 'dipteranVWHOInput'
}
\examples{
#Import the example input data for use with dipteranVWHO
data(dipteranVWHOInput)
dipteranVWHO(dipteranVWHOInput, evaluate=TRUE, percentileCut=5, lowCount=50)

}
\references{
Quinlan R, Smol JP (2001) Chironomid-based inference models for estimating end-of-summer hypolimnetic oxygen from south-central Ontario shield lakes. Freshwater Biology 46: 1529-1551

Quinlan R, Smol JP (2010) Use of Chaoborus subfossil mandibles in models for inferring past hypolimnetic oxygen. Journal of Paleolimnology 44: 43-50
}
\author{
Adam Jeziorski
}
