% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipteranVWHOInput.R
\docType{data}
\name{dipteranVWHOInput}
\alias{dipteranVWHOInput}
\title{An example input file for dipteranVWHO}
\format{
A data frame containing 154 observations of 7 variables.
}
\usage{
dipteranVWHOInput
}
\description{
This dataset demonstrates the required input format for \code{dipteranVWHO}

The \code{dipteranVWHOInput} data frame contains the following 4+ columns:
\itemize{
  \item Column 1: Label 1 (mainly tribe abbreviation)
  \item Column 2: Label 2 (mainly taxon name)
  \item Column 3: Label 3 (mainly the taxon code used for harmonization of the input data with the the 44 chironomid taxa included in the Quinlan and Smol 2010 model)
  \item Column 4+: Sample data
  }
}
