% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpDates.R
\name{interpDates}
\alias{interpDates}
\title{interpDates}
\usage{
interpDates(d, intervalWidth=0.5)
}
\arguments{
\item{d}{matrix or data frame containing the input data}

\item{intervalWidth}{(defaults to a constant value of 0.5): a single numeric values is used to indicate a constant sectioning resolution, but a vector (with interval widths for the full core) can also be used to provide the width of each interval in the core}
}
\value{
core, the table containing the interpolated dates for each interval.
}
\description{
\code{interpDates} interpolates dates for any undated intervals in a sediment core from the midpoint and age of the dated intervals, as well as the sectioning resolution.

Three sets of interpolated dates are returned:

\itemize{
  \item connectTheDotsDates: are calculated from straight line between sequential date pairs
  \item linearDates: are calculated by fitting a straight line through all dated intervals
  \item polynomialDates: are calculated by fitting a 2nd order polynomial line through all dated intervals
  }
  
For more involved approaches to the estimation of age-depth relationships consult Blaauw and Heegaard (2012).
}
\details{
the input data from the dated sediment intervals should be arranged in 2 columns:
\itemize{
  \item Column 1: midpoint depth of the dated interval
  \item Column 2: date of the interval
  }
}
\examples{
#4 dated intervals with a constant sectioning resolution of 0.5cm
interpDates.input <- cbind(c(0.25, 4.25, 8.25, 18.25), c(2017, 2000, 1950, 1850))
interpDates(interpDates.input)

#4 dated intervals in a core with slices 1-10 = 0.5cm, and 11-20 = 1.0cm
interpDates.input <- cbind(c(0.25, 4.25, 8.5, 13.5), c(2017, 1980, 1920, 1850))
intervalWidth <- c(rep(0.5,10), rep(1.0,10))
interpDates(interpDates.input, intervalWidth)
}
\references{
Blaauw M, Heegaard E (2012) 12. Estimation of age-depth relationships. In: Birks HJB, Lotter AF, Juggins S, Smol JP (eds.) Tracking Environmental Change Using Lake Sediments. Volume 5: Data Handling and Numerical Techniques. Springer, Netherlands, pp 379-413
}
\author{
Adam Jeziorski
}
