% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickGAM.R
\name{quickGAM}
\alias{quickGAM}
\title{quickGAM}
\usage{
quickGAM(x, y, xLabel = NULL, yLabel = NULL, yRange = NULL, skipAuto = FALSE)
}
\arguments{
\item{x}{vector containing the independent data/linear predictor}

\item{y}{vector containing the dependent data/response variable}

\item{xLabel}{(defaults to NULL): optional custom string to ensure consistent x-axis labels across the plots}

\item{yLabel}{(defaults to NULL): optional custom string to ensure consistent y-axis labels across the plots}

\item{yRange}{(defaults to NULL): optional custom range to ensure consistent y-axis across all plots}

\item{skipAuto}{(Defaults to FALSE): logical to indicate whether to skip the autocorrelation calculation and plot}
}
\description{
\code{quickGAM} is a function to fit a GAM and perform the related analyses described in Simpson (2018). This function essentially applies the code provided in that article's Supplementary Information, to a pair of vectors.
}
\details{
Two input vectors are required (x and y, the independent and dependent variable respectively). The optional arguments allow customization of the x and y axes labels to ensure they remain consistent across all graphs. Similarly the y-range can be manually defined.

The function produces six plots:
\itemize{
\item Basic scatterplot of the data
\item Estimated CAR1 process from the GAM fitted to the data ( i.e. autocorrelation check)
\item GAM-based trends fitted to the data
\item Estimated trends (w 20 random draws of the GAM fit to the data)
\item 95\% simultaneous confidence intervals (light grey) and across-the-function confidence intervals (dark grey) on the estimated trends.
\item Estimated first derivatives (black lines) and 95\% simultaneous confidence intervals of the GAM trends fitted to the data. Where the simultaneous interval does not include 0, the models detect significant temporal change in the response.
  }
}
\examples{
#Fit a GAM to the example chlorophyll a data in 'vrsChlaInput'
data(vrsChlaInput)
plotData <- vrsChla(vrsChlaInput)
#Values/labels to be preserved across all plots
indepVarLabel <- "Core Depth (cm)"
depenVarLabel <- expression(
 "VRS-Inferred Chlorophyll "~italic("a")~" (mg"\%.\%"g"^"-1"*textstyle(")"))
depenVarRange <- c(0, 0.0825)

#NOTE: drop=FALSE when reading in data columns preserves column names, allowing their use in plots.
quickGAM(plotData[,1, drop=FALSE],plotData[,2, drop=FALSE], xLabel = indepVarLabel, 
yLabel = depenVarLabel, yRange = depenVarRange)

}
\references{
Simpson GL (2018) Modelling Palaeoecological Time Series Using Generalised Additive Models. Frontiers in Ecology and Evolution [https://dx.doi.org/10.3389/fevo.2018.00149]
}
\author{
Adam Jeziorski
}
