% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrsChla.R
\name{vrsChla}
\alias{vrsChla}
\title{vrsChla}
\usage{
vrsChla(d, profilePlot = TRUE)
}
\arguments{
\item{d}{matrix or data frame containing the input data}

\item{profilePlot}{(defaults to TRUE): logical to control whether to produce the plot of sediment depth vs. inferred chlorophyll a values}
}
\value{
\code{chl.a.output} table of the inferred chlorophyll a values (mg/g dry mass) for each interval.
}
\description{
\code{vrsChla} infers chlorophyll a concentrations of sediments from spectral measurements of absorbance at wavelengths between 650-700 nm, following the approach described in Wolfe et al. (2006) and Michelutti et al. (2010).
}
\details{
The input data should be in the form of a matrix or data frame with 28 columns:
\itemize{
  \item Column 1: midpoint depth (cm) of the sediment interval
  \item Columns 2-27: values from the spectrophotometer for wavelengths 650-700 nm, measured every 2 nm
  }
NOTE: When using the Model 6500 series Rapid Content Analyzer at PEARL, the necessary values are contained in the 'spectra' tab of the excel file output (although they must be transposed). Ensure cells are formatted to 15 decimal places to avoid small rounding errors.
}
\examples{
#Infer the chlorophyll a values from the example spectral data
data(vrsChlaInput)
vrsChla(vrsChlaInput)

}
\references{
Michelutti N, Blais JM, Cumming BF, Paterson AM, Ruhland K, Wolfe AP, Smol JP (2010) Do spectrally inferred determinations of chlorophyll a reflect trends in lake trophic status? Journal of Paleolimnology 43: 208-217

Wolfe AP, Vinebrooke RD, Michelutti N, Rivard B, Das B (2006) Experimental calibration of lake-sediment spectral reflectance to chlorophyll a concentrations: methodology and paleolimnological validation. Journal of Paleolimnology 36: 91-100
}
\author{
Joshua Thienpont, Adam Jeziorski
}
