% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrsChlaInput.R
\docType{data}
\name{vrsChlaInput}
\alias{vrsChlaInput}
\title{An example input file for \code{vrsChla}}
\format{
a data frame containing 25 observations of 27 variables.
}
\usage{
data(vrsChlaInput)
}
\description{
This dataset demonstrates the required input format for \code{vrsChla} 
The \code{vrsChlaInput} data frame contains the following columns:
\itemize{
  \item Column 1: midpoint depth (cm) of the sediment interval
  \item Columns 2-27: values from the spectrophotometer for wavelengths 650-700 nm, measured every 2 nm
  }
}
\author{
Adam Jeziorski
}
