% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwhoQuinlan2010.R
\docType{data}
\name{vwhoQuinlan2010}
\alias{vwhoQuinlan2010}
\title{The Dipteran-Volume Weighted Hypolimnetic Oxygen (VWHO) Calibration Set from Quinlan and Smol 2010}
\format{
A data frame containing 54 observations of 47 variables.
}
\usage{
vwhoQuinlan2010
}
\description{
\code{vwhoQuinlan2010} is a 54x47 data frame containing the VWHO concentrations and sedimentary midge (chironomid and chaoborid) assemblage data for the 54 lake south-central Ontario calibration set described in Quinlan and Smol (2001, 2010).

The purpose of this dataset is to facilitate reproducing the described midge-inferred VWHO model for application to down-core midge sedimentary assemblage data.

Information on each lake (i.e. latitude, longitude, etc) and the 44 taxa included in the model are provided in Quinlan and Smol (2001,2010).

The \code{vwhoQuinlan2010} data frame contains the following 47 columns:
\itemize{
  \item Column 1: Lake name
  \item Column 2: Average VWHO values for each of the 54 lakes (mg/L)
  \item Column 3-46: Relative abundance data (as \% dipteran) for the 44 chironomid taxa included in the Quinlan and Smol 2010 model
  \item Column 47: Relative abundance data (as \% dipteran) for total chaoborids
  }
}
