% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiltseBrien.R
\name{wiltseBrien}
\alias{wiltseBrien}
\title{wiltseBrien}
\usage{
wiltseBrien(my.data, print=T, print.detail=T)
}
\arguments{
\item{my.data}{A data frame containing the variables to be tested. The variables should be arranged in columns with appropriate column names}

\item{print}{(Defaults to TRUE): logical to indicate whether to display the degrees of freedom, chi-squared, and p-values for the grand mean, main effects, interactions, and equal correlations}

\item{print.detail}{(Defaults to TRUE): logical to indicate whether to display the variables included in the test and the correlation coefficient for the grand mean}
}
\value{
\code{output} table of results for use with other functions
}
\description{
\code{wiltseBrien} performs a Brien's test on a correlation matrix to test whether changes in the underlying variables are coherent. 

This function is detailed in Brendan Wiltse's PhD thesis and calculates the grand mean, main effects interactions and equal correlation for the correlation matrix and returns the degrees of freedom, chi-squared value and p-value for each test.
}
\details{
\code{wiltseBrien} Full details on the Brien's test and the underlying mathematics can be found in Brien et al. 1984
}
\examples{
#Import the example input data for use with Brendan Wiltse's functions
data(wiltseInput)
wiltseBrien(wiltseInput[,2:9])
}
\references{
Brien CJ, Venables WN, James AT, Mayo O (1984) An analysis of correlation matrices: equal correlations. Biometrika 71: 545-54

Wiltse B (2014) The response of Discostella species to climate change at the Experimental Lakes Area, Canada. PhD Thesis Queen's University
}
\author{
Brendan Wiltse
}
