% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiltseInput.R
\docType{data}
\name{wiltseInput}
\alias{wiltseInput}
\title{An example input file for \code{wiltseBrien} and \code{wiltseDecompose}}
\format{
a data frame containing 12 observations of 9 variables.
}
\usage{
wiltseInput
}
\description{
This dataset is used to demonstrate the use of the two functions included in Brendan Wiltse's PhD thesis \code{wiltseBrien} and \code{wiltseDecompose}
\itemize{
  \item Column 1: Date of the interval
  \item Columns 2-9: PCA axis-1 scores for the dominant diatom species in 8 lakes
  }
}
\author{
Brendan Wiltse
}
